[GB.FORM] * BUG: FileChooser: Fix an infinite loop when trying to initialize the path   with an unknown file.
[GB.MEDIA.FORM]
* NEW: New component based on gb.media that includes a full media player 
  widget with controls.

[GB.NET.SMTP]
* NEW: SmtpClient is now a virtual control.


git-svn-id: svn://localhost/gambas/trunk@6408 867c0c6c-44f3-4631-809d-bfa615b0a4ec

+++ gb.media/src/c_media.c
@@ -1617,12 +1617,12 @@ BEGIN_PROPERTY(MediaPipeline_Position)
 	}
 	else
 	{
-		guint64 pos = VPROP(GB_FLOAT) * 1E9;
+		gint64 pos = VPROP(GB_FLOAT) * 1E9;
 		
 		if (pos < 0) 
 			pos = 0;
-		
-		gst_element_seek_simple(ELEMENT, GST_FORMAT_TIME, GST_SEEK_FLAG_FLUSH | GST_SEEK_FLAG_KEY_UNIT, pos);
+
+		gst_element_seek_simple(ELEMENT, GST_FORMAT_TIME, GST_SEEK_FLAG_FLUSH | GST_SEEK_FLAG_KEY_UNIT, (guint64)pos);
 	}
 
 END_PROPERTY
